Complete Flex Exercise 02
This one warrants further explanation. Centering the flex items inside
the header div was simple enough with justify-content and align-items
both being set to center. Having the logo be centered was also easy,
by setting its left and right margins to auto.

The tricky parts came with the link lists. First, while I tried
removing margins from the ul elements, I still had large spacing on the
left side of the left links. This was apparently due to padding inherent
to ul/ol/li. By setting margin and padding of the ul to 0, this is
resolved.

The bit I had the hardest time with was creating a gap between each li.
Initially, I tried making each ul and left-/right-links flex containers
and justifying contents to space-around. This didn't work, and once I
decided to have the header have a minimum padding of 10px, I tried
space-between which still didn't work. I believe this was because no
empty space was available with the logo having auto margins, forcing
the link lists to a specific size with no empty space. I tried
commenting out my left-/right-link selectors and noticed the format
was largely still acceptable, and applied Occam's Razor and deleted
them after that. I decided to apply Occam's Razor once more, and just
apply a 20px column-gap (because the gaps in the example seemed larger than
the margins on the side) to the ul selector, which worked great and
meant I had to create fewer selectors myself.